Feature point details dialog

+++ app/src/main/java/com/example/gh_coursework/domain/usecase/point_details/GetPointDetailsUseCaseImpl.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 
 class GetPointDetailsUseCaseImpl(private val repository: TravelRepository) :
     GetPointDetailsUseCase {
-    override fun invoke(pointId: Int): Flow<PointDetailsDomain?> {
+    override fun invoke(pointId: Long): Flow<PointDetailsDomain?> {
        return repository.getPointOfInterestDetails(pointId)
     }
 }
