+++ app/src/main/java/com/example/gh_coursework/ui/private_point/PointViewModel.kt
@@ -31,7 +31,7 @@ class PointViewModel(
         }
     }
 
-    fun getPointDetailsPreview(pointId: Int): Flow<PrivatePointDetailsPreviewModel?> {
+    fun getPointDetailsPreview(pointId: Long): Flow<PrivatePointDetailsPreviewModel?> {
         return getPointDetailsUseCase.invoke(pointId).map { details -> mapPointDetailsDomainToModel(details) }
     }
